experts/requirements.txt
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.1.0
fastapi==0.88.0
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-bigquery==3.4.1
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
greenlet==2.0.1
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
packaging==21.3
passlib==1.7.4
pika==1.3.1
pluggy==1.0.0
proto-plus==1.22.1
protobuf==4.21.12
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
redis==4.5.3
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
